# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText - OpenAPI 3.0
  description: |
    This API allows sharing and receiving messages among users in WasaText.
  version: 0.0.1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      title: User
      type: object
      description: Details of a registered user
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Unique identifier of the User
        username:
          type: string
          description: Username of the User
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_]+$"
        photo:
          type: string
          format: uri
          pattern: "^[a-zA-Z0-9_]+$"
          description: URL to the User's profile photo
          minLength: 1
          maxLength: 20
      required:
        - id
        - username
...
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              description: some description
              type: object
              properties:
                name:
                  type: string
                  description: Name of the user
                  example: Maria
                  pattern: '^[a-zA-Z0-9]+$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                description: some description
                type: object
                properties:
                  identifier:
                    type: string
                    pattern: "^[a-zA-Z0-9_]+$"
                    minLength: 3
                    maxLength: 50
                    description: Identifier of the logged-in user
                    example: "abcdef012345"

  # I Send a JSON with name of user, server finds orr not, if not find create an user, if finds back token. 
  /users/me/username:
    put:
      tags:
        - Users
      summary: Set my username
      description: Updates the username of the authenticated user.
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload for updating the username
              properties:
                username:
                  type: string
                  description: New username for the user
                  minLength: 3
                  maxLength: 50
                  pattern: "^[a-zA-Z0-9_]+$"
                  example: "new_username"
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Response object for a successful username update
                properties:
                  message:
                    type: string
                    description: Success message confirming the username update
                    pattern: ^.*$
                    minLength: 3
                    maxLength: 50
                    example: "Username updated successfully"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: string
                description: Error message for invalid input
                pattern: "^[a-zA-Z0-9_]+$"
                minLength: 3
                maxLength: 50
        '409':
          description: Username already in use
          content:
            application/json:
              schema:
                type: string
                description: Error message indicating the username is already taken
                pattern: "^[a-zA-Z0-9_]+$"
                minLength: 3
                maxLength: 50

# server checks format and unuque of username, update and back 

  #  новый ендпоинт
  /users/me/photo:
      put:
        tags:
          - Users
        summary: Set my photo
        description: Updates the photo of the authenticated user.
        operationId: setMyPhoto
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                # ПОМЕНЯТЬ
                description: Payload containing the URL of the new user photo.
                properties:
                  photo:
                    type: string
                    description: The URL of the new photo to set as the user's profile photo.
                    pattern: "^(https?://).+$" 
                    minLength: 1
                    maxLength: 100
                    format: uri
                    example: "https://example.com/photos/user-photo.png"
        responses:
          '200':
            description: User photo updated successfully.
            content:
              application/json:
                schema:
                  type: object
                  description: Success response containing a confirmation message.
                  properties:
                    message:
                      type: string
                      description: A confirmation message indicating the user photo was updated successfully.
                      pattern: ^.*$
                      minLength: 3
                      maxLength: 100
                      example: "User photo updated successfully."
          '404':
            description: User not found.
            content:
              application/json:
                schema:
                  type: string
                  description: Error message indicating the specified user could not be found.
                  pattern: ^.*$
                  minLength: 3
                  maxLength: 50
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user details
      description: Retrieves the details of a user by their unique identifier.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The unique identifier of the user.
      responses:
        '200':
          description: User details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: string
                description: Error message indicating the user could not be found.
                pattern: ^.*$
                minLength: 3
                maxLength: 50






# doLogin (see simplified login)
#   POST /session
#   Body (JSON): { "name": "<username>" }
#   -> create user if not exists / return existing user
#   <- 201 { "identifier": "<token/userID>" } (store in localStorage)

# setMyUserName
#   PUT /users/me/username
#   Body (JSON): { "username": "<new_username>" }
#   -> validate + ensure unique + update username
#   <- 200 { "message": "Username updated successfully" }

# getMyConversations
#   GET /users/{id}/conversations
#   -> return list of user conversations (private/group, last activity, participant info)
#   <- 200 [ Conversation, ... ]

# getConversation
#   GET /conversations/{c_id}
#   -> return conversation details + all messages (including reply_to info if present)
#   <- 200 { "conversation": {...}, "messages": [...] }

# sendMessage
#   POST /conversations/{conversation_id}/messages
#   Body (multipart/form-data):
#     - text: content + content_type="text"
#     - file: file=<blob> (backend infers photo/gif by extension)
#     - optional: reply_to=<message_id>
#   -> check membership, save upload to /uploads if file, insert message into DB
#   <- 201 { message, content_type, content, sender_username, sender_photo }

# forwardMessage
#   POST /conversations/{conversation_id}/messages/{message_id}/forward/{target_conversation_id}
#   Target:
#     - target_conversation_id = number -> forward to existing chat/group (body can be {})
#     - target_conversation_id = "new"  -> Body (JSON): { "target_username": "<user_or_group_name>" }
#   -> check membership source+target; if "new" find/create target conversation;
#      read original message content; insert new message into target as current user
#   <- 200 { "message": "Message forwarded successfully" }

# commentMessage
#   POST /conversations/{conversation_id}/messages/{message_id}/comments
#   Body:
#     - file: multipart/form-data { file=<blob> } (photo/gif)
#     - text/emoji: JSON { "content_type": "text"|"emoji", "content": "..." }
#   -> check membership; save upload if file;
#      if original message deleted -> add as normal message; else -> add comment linked to message
#   <- 201 { message, content_type, content } (or "comment added as normal message")

# uncommentMessage
#   DELETE /conversations/{conversation_id}/messages/{message_id}/comments/{comment_id}
#   -> check membership + comment owner; delete comment
#   <- 200/204 success

# deleteMessage
#   DELETE /conversations/{conversation_id}/messages/{message_id}
#   -> verify message exists; convert its comments to normal messages; delete message
#   <- 200/204 success

# addToGroup
#   POST /groups/{c_id}/members
#   Body (JSON): { "usernames": ["u1","u2", ...] }   (backend expects usernames list)
#   -> requester must be member; add each user to convmembers if not already in group
#   <- 200 { message, c_id, added_users: [...] }

# leaveGroup
#   DELETE /groups/{c_id}/leave
#   -> check it is a group + membership; remove user from group;
#      if group becomes empty -> optionally delete group
#   <- 200 { "message": "Successfully left the group" }

# setGroupName
#   PUT /groups/{c_id}/name
#   Body (JSON): { "new_name": "<name>" }   (backend uses new_name)
#   -> check group + membership + name uniqueness; update group name
#   <- 200 { message, group_id, new_name }

# setMyPhoto
#   PUT /users/me/photo
#   Spec says URL JSON: { "photo": "https://..." }
#   (if you implement upload instead: multipart/form-data { photo=<file> })
#   -> update user's photo (store URL or saved /uploads path)
#   <- 200 { message, photo? }

# setGroupPhoto
#   PUT /conversations/{c_id}/set-group-photo
#   Body (multipart/form-data): { photo=<file> }  (backend expects "photo")
#   -> check group + membership; save file to /uploads; update group photo in DB
#   <- 200 { message, photo: "/uploads/..." }